+++ src/main/java/com/aetherteam/aether/item/food/GummySwetItem.java
@@ -1,7 +1,6 @@
 package com.aetherteam.aether.item.food;
 
 import com.aetherteam.aether.AetherConfig;
-import com.aetherteam.aether.item.AetherItems;
 import com.aetherteam.aether.item.miscellaneous.ConsumableItem;
 import net.minecraft.world.InteractionHand;
 import net.minecraft.world.InteractionResult;
@@ -14,8 +13,8 @@ import net.minecraft.world.item.ItemUseAnimation;
 import net.minecraft.world.level.Level;
 
 public class GummySwetItem extends Item implements ConsumableItem {
-    public GummySwetItem() {
-        super(new Item.Properties().rarity(AetherItems.AETHER_LOOT).food(AetherFoods.GUMMY_SWET));
+    public GummySwetItem(Properties properties) {
+        super(properties);
     }
 
     /**