+++ src/main/java/com/aetherteam/aether/item/food/HealingStoneItem.java
@@ -5,12 +5,11 @@ import net.minecraft.world.effect.MobEffects;
 import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
-import net.minecraft.world.item.Rarity;
 import net.minecraft.world.level.Level;
 
 public class HealingStoneItem extends Item {
-    public HealingStoneItem() {
-        super(new Item.Properties().rarity(Rarity.RARE).food(AetherFoods.HEALING_STONE));
+    public HealingStoneItem(Properties properties) {
+        super(properties);
     }
 
     /**